# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoJigGenerator API",
    "version": "baseline"
  },
  "paths": {
...
    "/api/works/{work_id}/promote/part": {
      "post": {
        "tags": [
          "works"
        ],
        "summary": "Promote Part",
        "description": "현재 형상 버전을 부품 카탈로그에 올린다.",
        "operationId": "promote_part_api_works__work_id__promote_part_post",
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromotePartRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/parts": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "List Parts",
        "operationId": "list_parts_api_parts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PartSummaryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/parts/{part_id}": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "Get Part",
        "operationId": "get_part_api_parts__part_id__get",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Part Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "parts"
        ],
        "summary": "Update Part",
        "operationId": "update_part_api_parts__part_id__patch",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Part Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "parts"
        ],
        "summary": "Delete Part",
        "operationId": "delete_part_api_parts__part_id__delete",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Part Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/parts/{part_id}/versions": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "List Versions",
        "operationId": "list_versions_api_parts__part_id__versions_get",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Part Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartVersionOut"
                  },
                  "title": "Response List Versions Api Parts  Part Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/parts/{part_id}/copy-to-work": {
      "post": {
        "tags": [
          "parts"
        ],
        "summary": "Copy To Work",
        "description": "부품의 레시피로 **내 작업**을 새로 만든다. 남의 부품을 고치는 길은 이것뿐이다.",
        "operationId": "copy_to_work_api_parts__part_id__copy_to_work_post",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Part Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyToWorkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "mine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "내가 건 것만",
              "default": true,
              "title": "Mine"
            },
            "description": "내가 건 것만"
          },
          {
            "name": "work_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(queued|running|done|failed)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "화면이 도는 동안 이것을 폴링한다 — `progress` 가 단계마다 늘어난다.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artifacts/{artifact_id}/download": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Download Artifact",
        "operationId": "download_artifact_api_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CopyToWorkRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          }
        },
        "type": "object",
        "title": "CopyToWorkRequest",
        "description": "부품(버전)의 레시피로 내 작업을 새로 만든다 — 「내 공간으로 복사」."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "requested_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Id"
          },
          "requested_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Name"
          },
          "work_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Id"
          },
          "work_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Name"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          },
          "progress": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "type": "array",
            "title": "Progress"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactOut"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "requested_by_id",
          "requested_by_name",
          "work_id",
          "work_name",
          "options",
          "progress",
          "summary",
          "error",
          "artifacts",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobOut"
      },
...
      "Page_JobOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[JobOut]"
      },
      "Page_PartSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PartSummaryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PartSummaryOut]"
      },
...
      "PartOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "work_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Id"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count"
          },
          "current": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartVersionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "jig_count": {
            "type": "integer",
            "title": "Jig Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "owner_id",
          "owner_name",
          "work_id",
          "current_version",
          "version_count",
          "current",
          "jig_count",
          "created_at",
          "updated_at"
        ],
        "title": "PartOut"
      },
...
      "PartUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "PartUpdateRequest"
      },
      "PartVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "part_id": {
            "type": "string",
            "format": "uuid",
            "title": "Part Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "recipe": {
            "additionalProperties": true,
            "type": "object",
            "title": "Recipe"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "promoted_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted By Id"
          },
          "promoted_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted By Name"
          },
          "work_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "part_id",
          "number",
          "recipe",
          "job",
          "note",
          "promoted_by_id",
          "promoted_by_name",
          "work_version_id",
          "created_at"
        ],
        "title": "PartVersionOut"
      },
...
      "PromotePartRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "title": "PromotePartRequest",
        "description": "현재 버전을 부품으로. 이 작업에서 이미 승격한 부품이 있으면 거기에 다음 버전으로\n붙는다."
      },
...
      "WorkOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count"
          },
          "current": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "jig_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Jig Options"
          },
          "jig_run_count": {
            "type": "integer",
            "title": "Jig Run Count"
          },
          "last_jig_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Jig Status"
          },
          "promoted_part_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted Part Id"
          },
          "promoted_jig_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoted Jig Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "owner_id",
          "owner_name",
          "current_version",
          "version_count",
          "current",
          "jig_options",
          "jig_run_count",
          "last_jig_status",
          "promoted_part_id",
          "promoted_jig_id",
          "created_at",
          "updated_at"
        ],
        "title": "WorkOut"
      },
...
    }
  }
}